requirements.txt
openai>=1.0.0
colorama>=0.4.4
pyyaml>=6.0
rich>=12.0.0
click>=8.0.0
tqdm>=4.62.0
bandit>=1.7.0
safety>=2.0.0
gitpython>=3.1.0
tabulate>=0.8.9
requests>=2.27.0
python-dotenv>=0.19.0
